requirements.txt
gspread==6.1.3
google-auth==2.40.3
pandas==2.2.3
plotly==5.22.0
streamlit==1.32.0
